+++ lib/features/auth/data/repositories/auth_repository_firebase.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:kwadra/features/auth/data/data_sources/firebase_auth_data_source.dart';
 
 import '../../../../core/error/failure.dart';
-import '../../domain/entities/sign_in_response.dart';
+import '../../domain/entities/auth_response.dart';
 import '../../domain/repositories/auth_repository.dart';
 
 class AuthRepositoryFirebase implements AuthRepository {
@@ -12,7 +12,7 @@ class AuthRepositoryFirebase implements AuthRepository {
   AuthRepositoryFirebase(this.firebaseAuthDataSource);
 
   @override
-  Future<Either<Failure, SignInResponse>> signInWithEmailAndPassword(
+  Future<Either<Failure, AuthResponse>> signInWithEmailAndPassword(
       String email, String password) async {
     try {
       if (password.length < 8) {
@@ -26,7 +26,28 @@ class AuthRepositoryFirebase implements AuthRepository {
       final user = await firebaseAuthDataSource.signInWithEmailAndPassword(
           email, password);
       if (user == null) return const Left(ParamsFailure("Sign in failed"));
-      return const Right(SignInResponse('Sign in success'));
+      return const Right(AuthResponse('Sign in success'));
+    } catch (error) {
+      return Left(ParamsFailure(error.toString()));
+    }
+  }
+
+  @override
+  Future<Either<Failure, AuthResponse>> signUpWithEmailAndPassword(
+      String email, String password) async {
+    try {
+      if (password.length < 8) {
+        return const Left(ParamsFailure("Invalid password length"));
+      }
+
+      if (!_isValidEmail(email)) {
+        return const Left(ParamsFailure("Invalid email format"));
+      }
+
+      final user = await firebaseAuthDataSource.signUpWithEmailAndPassword(
+          email, password);
+      if (user == null) return const Left(ParamsFailure("Sign up failed"));
+      return const Right(AuthResponse('Sign up success'));
     } catch (error) {
       return Left(ParamsFailure(error.toString()));
     }