+++ lib/features/auth/domain/repositories/auth_repository.dart
@@ -1,9 +1,12 @@
 import 'package:dartz/dartz.dart';
 
 import '../../../../core/error/failure.dart';
-import '../entities/sign_in_response.dart';
+import '../entities/auth_response.dart';
 
 abstract class AuthRepository {
-  Future<Either<Failure, SignInResponse>> signInWithEmailAndPassword(
+  Future<Either<Failure, AuthResponse>> signInWithEmailAndPassword(
+      String email, String password);
+
+  Future<Either<Failure, AuthResponse>> signUpWithEmailAndPassword(
       String email, String password);
 }