+++ lib/features/auth/domain/usecases/sign_up_with_email_and_password.dart
@@ -4,12 +4,12 @@ import 'package:kwadra/features/auth/domain/repositories/auth_repository.dart';
 import '../../../../core/error/failure.dart';
 import '../entities/auth_response.dart';
 
-class SignInWithEmailAndPassword {
+class SignUpWithEmailAndPassword {
   final AuthRepository _authRepository;
 
-  SignInWithEmailAndPassword(this._authRepository);
+  SignUpWithEmailAndPassword(this._authRepository);
 
   Future<Either<Failure, AuthResponse>> call(String email, String password) {
-    return _authRepository.signInWithEmailAndPassword(email, password);
+    return _authRepository.signUpWithEmailAndPassword(email, password);
   }
 }