+++ lib/features/auth/presentation/bloc/auth_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:kwadra/features/auth/domain/entities/sign_in_response.dart';
 
+import '../../domain/entities/auth_response.dart';
 import '../../domain/repositories/auth_repository.dart';
 
 part 'auth_event.dart';
@@ -12,6 +12,7 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
 
   AuthBloc({required this.authRepository}) : super(AuthInitial()) {
     on<SignInRequested>(_onSignInRequested);
+    on<SignUpRequested>(_onSignUpRequested);
   }
 
   Future<void> _onSignInRequested(
@@ -20,6 +21,17 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
   ) async {
     emit(const AuthLoading());
     final failureOrSuccess = await authRepository.signInWithEmailAndPassword(
+        event.email, event.password);
+    emit(failureOrSuccess.fold(
+      (failure) => AuthFailure(failure.message),
+      (response) => AuthSuccess(response),
+    ));
+  }
+
+  Future<void> _onSignUpRequested(
+      SignUpRequested event, Emitter<AuthState> emit) async {
+    emit(const AuthLoading());
+    final failureOrSuccess = await authRepository.signUpWithEmailAndPassword(
         event.email, event.password);
     emit(failureOrSuccess.fold(
       (failure) => AuthFailure(failure.message),