+++ lib/features/auth/presentation/bloc/auth_event.dart
@@ -16,3 +16,13 @@ class SignInRequested extends AuthEvent {
   @override
   List<Object> get props => [email, password];
 }
+
+class SignUpRequested extends AuthEvent {
+  final String email;
+  final String password;
+
+  const SignUpRequested(this.email, this.password);
+
+  @override
+  List<Object> get props => [email, password];
+}