+++ test/features/auth/data/data_sources/firebase_auth_data_source_test.dart
@@ -27,6 +27,7 @@ void main() {
       dataSource = FirebaseAuthDataSourceImpl(firebaseAuth: mockFirebaseAuth);
     });
 
+    //// SIGN IN ////
     test('should return success when the response code is 200', () async {
       // arrange
       when(() => mockFirebaseAuth.signInWithEmailAndPassword(
@@ -61,6 +62,44 @@ void main() {
 
       // assert
       verify(() => mockFirebaseAuth.signInWithEmailAndPassword(
+          email: 'valid@mail.com', password: 'password')).called(1);
+    });
+
+    //// SIGN UP ////
+    test('should return success when the response code is 200', () async {
+      // arrange
+      when(() => mockFirebaseAuth.createUserWithEmailAndPassword(
+          email: 'valid@mail.com',
+          password: 'password')).thenAnswer((_) async => mockUserCredential);
+
+      when(() => mockUserCredential.user).thenReturn(mockUser);
+
+      // act
+      final result = await dataSource.signUpWithEmailAndPassword(
+          'valid@mail.com', 'password');
+
+      // assert
+      expect(result, mockUser);
+      verify(() => mockFirebaseAuth.createUserWithEmailAndPassword(
+          email: 'valid@mail.com', password: 'password')).called(1);
+    });
+
+    test('should throw an exception when the response code is not 200',
+        () async {
+      // arrange
+      when(() => mockFirebaseAuth.createUserWithEmailAndPassword(
+              email: 'valid@mail.com', password: 'password'))
+          .thenThrow(FirebaseAuthException(code: 'wrong-password'));
+
+      // act
+      expect(
+        () async => await dataSource.signUpWithEmailAndPassword(
+            'valid@mail.com', 'password'),
+        throwsA(isA<FirebaseAuthException>()),
+      );
+
+      // assert
+      verify(() => mockFirebaseAuth.createUserWithEmailAndPassword(
           email: 'valid@mail.com', password: 'password')).called(1);
     });
   });